+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed the storage of the market data in the data provider service to store only the newly fetched quotes
+- Fixed the immediate expiration of a portfolio snapshot with errors
 - Fixed the missing country mapping of _Congo (Dem. Rep. of the)_ and _Congo (Rep. of)_ in the _Financial Modeling Prep_ service
 
 ## 3.66.0 - 2026-09-03